requirements.txt
aniso8601==9.0.1
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
distlib==0.3.4
dnspython==2.2.1
filelock==3.7.1
Flask==2.1.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
mailjet-rest==1.3.4
MarkupSafe==2.1.1
numpy==1.22.4
pandas==1.4.2
passlib==1.7.4
platformdirs==2.5.2
pycodestyle==2.9.1
PyJWT==2.4.0
pymongo==4.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-http-client==3.3.7
pytz==2022.1
requests==2.28.1
scikit-learn==1.1.1
scipy==1.8.1
sendgrid==6.9.7
six==1.16.0
sklearn==0.0
starkbank-ecdsa==2.0.3
threadpoolctl==3.1.0
toml==0.10.2
urllib3==1.26.12
virtualenv==20.14.1
Werkzeug==2.1.2
